+++ R/skip.R
@@ -14,7 +14,7 @@
 skip <- function(..., w = FALSE){
 
   # there may be no need to check
-  has <- lapply(list(...), is.err)
+  has <- lapply(list(...), is.problem)
   has <- unlist(has)
 
   if(!any(has))