+++ tests/testthat/test_is.R
@@ -4,10 +4,10 @@ test_that("Is tests", {
 
   testthat::expect_true(is.e(error))
   testthat::expect_true(is.w(warning))
-  testthat::expect_true(is.err(error))
-  testthat::expect_true(is.err(warning))
+  testthat::expect_true(is.problem(error))
+  testthat::expect_true(is.problem(warning))
 
   testthat::expect_false(is.e("hello"))
   testthat::expect_false(is.w("world"))
-  testthat::expect_false(is.err("world"))
+  testthat::expect_false(is.problem("world"))
 })